// typescript-projects/utility-backups/src20250516BeforeErrorHandlingChange/interface/types.ts
export type Flatten<T> = T extends Record<string, unknown>
  ? UnionToIntersection<
      {
        [K in keyof DeepRequired<T>]: K extends string
          ? T[K] extends Record<string, unknown>
            ? {
                [P in `${K}.${keyof Flatten<T[K]> & string}`]: Flatten<
                  T[K]
                >[P extends `${K}.${infer R}` ? R : never];
              } & { [P in K]: T[K] }
            : { [P in K]: T[K] }
          : never;
      }[keyof T]
    >
  : never;

export type UnionToIntersection<T> = (
  T extends unknown ? (x: T) => void : never
) extends (x: infer R) => void
  ? R
  : never;

/** Recursively make all properties of a map/object required */
export type DeepRequired<T> = {
  [K in keyof T]-?: Exclude<T[K], undefined> extends Record<string, unknown>
    ? DeepRequired<T[K]>
    : T[K];
};
// ...
export type TUtilityMeta = TEbillsUtilityMeta;
export type TEbillsUtilityMeta = {
  // /** Unique identifier for the customer, at the merchant */
  // customerId: string;
  // merchantId: UtilityMerchant;
  // productId: UtilityProduct;
  // planId: string;
  // amount: number;
  /** Provider Identifier. E.g. ebills */
  provider: ProviderId;
  /** E.g. mtn, mtn-n. Provider-specific */
  merchant: string;
  /** Product ID at provider */
  variationId?: string;
  /** Customer ID at merchant for airtime and internet products */
  phoneNumber?: string;
  /** Customer ID at merchant for bet, prepaid, and postpaid products */
  customerId?: string;
  /** Customer ID at merchant for cableTV product */
  cardNumber?: string;
};

export type TUtilityWithdrawalReadData = {
  adminInfo: {
    /** Path to the related naira_admins doc */
    naira_path: string;
  };
  amount?: number;
  /** doc ID for this, and the related naira_admins and naira_metadata update docs */
  hash: string;
  userId: string;
  // reason: Utility;
  reasonMeta: TUtilityMeta;
};